terraform_agent/agent/requirements.txt
langgraph>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
